Refactor Cinema 4D addon to align with AYON Blender structure
- Moved `Cinema4DHost` class from `api/pipeline.py` to `api/host.py` to follow standard AYON addon structure.
- Updated `api/__init__.py` to expose `Cinema4DHost` from the new location.
- Updated `api/host.py` to handle plugin path registration (`PUBLISH_PATH`, `LOAD_PATH`, etc.) and delegating workfile operations to `workio.py`.
- Cleaned up `api/pipeline.py` to retain only pipeline-specific utilities (`iter_containers`, `containerise`, `on_task_changed`).
- Ensured import compatibility in `startup/ayon_cinema4d.pyp`.
- Verified no broken references remain.
